Remove mockserver-junit-jupiter
